Improve formatting in details for data that's a list
Fixes #29

Not using DT because for some reason it doesn't want to show up. If I include
the table as a DT table and open the details drawer, I see nothing *until* I
open the chrome developer tools. Then it shows up, but is too wide (and altering
fillContainer doesn't seem to fix things satisfactorily). Will stick with a
static table for now and if we need the DT functionality later we can revisit.