Check for complete columns (#97)
* Team alteration

Altered the team number and folder to upload to for testing.

* Fixing conflict with master

Team membership was updated in master, but I was working on a different branch before creating this branch. When I brought the changes with me, it had a conflict with the master due to the changing of the team and project ids with the new format.

* Added initial check-cols-complete.R and tests

* Removed leftover from conflict

Not sure why this section stuck around after the conflict resolution. Removing it since it's unnecessary.

* Updated example

* Stopped dropping of dataframe when one column or value

* Linting & styler changes

* Edited behavior to correctly list the columns that should be complete

Takes in the `required_cols` list and prints it in the message saying they should be complete.

* Changed `strict` default to `TRUE`

* Updated tests for having the new `strict` default as `TRUE`

* Update comment R/check-cols-complete.R for clarity

Co-Authored-By: Kara Woo <karawoo@users.noreply.github.com>

* Update description of required_cols param of check_cols_complete() for clarity

* Update documentation

Ran `devtools::documentation()` to update documentation.