Enforce that something has been entered as the study name
If a user answers "no" to the "Does the study currently exist?" question and doesn't input anything into the textbox, then it will upload the file with an empty string as the study name. Added check to verify that there is more than an empty string before the submit button will function.