Add template for validating complete columns
Created base template for validating that required columns are complete. Based off of empty column validation. Still need to put in the columns that are required for each, but not sure what those are, yet. Required columns could be different based on species, as well, in which case this would need to be accounted for.