Documentation module (#93)
* Team alteration

Altered the team number and folder to upload to for testing.

* Added UI elements and most functionality

Changed UI to have the sidebar with tabs for the validator and documentation. Added the UI for the documentation uploading (outside of having the correct studies populate the selectionlist). Have functionality working.

* Linting

* Added functionality for documents section and linted

Added:
- populated studies in the drop down selection box.
- put a tryCatch around the synapse uploading in case it fails.
- added a small notification when files are uploaded; hopefully this will get updated to something better in the future.
- changed `for` loop to `purrr:map2`.
- updated `save_to_synapse()` to have an annotations parameter that is saved to synapse with the file.

* roxygen2 updated the namespace file

* Added the indications to document submit button

* Changed team & folder ids back to original

* Linting and moving get_studies() to own file

* Prepend study name and lint

Before uploading, prepend the study name to the document name.

* Add documentation for button-indicator-helpers.R

* Linting

* Added instructions

Instructions include description of documents to upload.

* Made the panel for getting studies into a module

* Made documentation content and button indicator into a module

* Changed the modules directory casing to match others

The other directories have lowercase names, except for R. Changed Modules to modules to match.

* Midway through fix to make Modules dir to be lowercase.

Can't just make the directory lowercase and then commit. Have to do a convoluted workflow to make this work. First step: created a temp directory for name change.

* Finishing directory name change

Second step of name change: change to desired name.

* Fixed a module error and added some comments

Needed to match the id sent to the module with the tabItem name in upload-documentation.R. This makes it so you can have multiple instances without namespace clashes.

* Optimize file upload in upload-documentation.R

Had a series of if-else statements depending on the number of files or file type. Instead, combine the files into a single dataframe, gather the info needed, and use purrr map function to iteration over.

* Put validator ui back in main function call, updated instructions

Updated instructions to remove the template for assay descriptions, and bolded the items that should be in the description. Moved the UI for the validator into the main function body.

* Fiddled with spacing

* Linting & removed unnecessary source calls

For some reason, I have to use the source calls if I'm not in the dccvalidator working directory. The functions for the portions I've added simply don't load. However, it works fine if your working directory is dccvalidator.

* Fixed spacing

* Styler changes

* Move mods to R folder and append name with mod

Modules were in a subdirectory 'modules' in the 'R' directory. Moved these to the 'R' directory and appended 'mod-' to the front of each name.

* Updated documentation

Ran 'devtools::document()' to update package documentation.

* Style changes

* Move css style for button to css stylesheet

* Updated documentation

Changed the references to the 'shiny::callModule' function in the documentation to '[shiny::callModule()]'.

* Removed unncessary reference to stylesheet

* Added study table id parameter to get_studies() and updated parameter chain

'get_studies()' now takes the synId of the consortiums study table for getting the list of studies. The synId is hard-coded in app-ui.R and passed to the various modules until it gets to 'get_studies()'.

* Incremented version and added self as an author

* Updates from styler, devtools::documentation(), and lintr

Also removed an extra comma in DESCRIPTION that I had accidentally put in.

* Added missed parent_folder param documentation

Reran the documentation tool after adding.

* Get session from `getDefaultReactiveDomain()`

Originally was using a non-exported function from shinyjs (`getSession()`), but it was pointed out that it wrapped an exported shiny function (`getDefaultReactiveDomian()`).

* Required team membership to load study names and enable document module inputs

Updates:
- Moved `callModule()` for documentation server to the `if'-statment that checks for team membership and creates the folder.
- Made it so the inputs in documentation module are disabled until team membership has been checked and there is the initial call to `upload_documents_server()`.
- Moved the `study_table_id` to the server call since this will ensure the user has team membership.
- Ran documentation tool to update documents.

* styler changes

Only mod-upload-documentation.R and mod-get-study-names.R were edited.

* Lint change

* Enforce that something has been entered as the study name

If a user answers "no" to the "Does the study currently exist?" question and doesn't input anything into the textbox, then it will upload the file with an empty string as the study name. Added check to verify that there is more than an empty string before the submit button will function.

* Change layout of documentation module

* Increment version to be ahead of master

* Remove comma

* remove comma

* Add my team

* Update dependencies

* Replace my team id and folder with original