Custom file uploads by type (#103)
* Team alteration

Altered the team number and folder to upload to for testing.

* Rearrange question items and file upload items

* Update species options

* Made individual metadata upload only appear if species is not drosophila

* Changed species options & updated check_cols_individual

Previous options were ` human`, `animal`. Updated to `human`, `drosophila`, `mouse and other animal models`. For the mouse & other models, I considered those the `general` category. Updated check_cols_individual() to have template options for each species option. However, the `general` option is currently the previous template for `animal` and I put `drosophila` under that same template for now.

* Remove drosophila option from check_cols_individual

There should not be individual metadata for drosophila.

* Add storage of files in a `reactiveValue`

Cannot change the value of an input from UI, but can store as a reactiveValue that can be changed. Doing this allows for resetting data to NULL, without being affected by the lingering file information in the input variable.

* Lint the species list, remove extra comma

* Add event for drosophila that resets the individual fileInput and file data

Resetting `fileInput` does not reset the stored input data from the `fileInput`. Instead, have to store separately and set to `NULL` separately. This propogates down, updating the reactive variables.

* Use template options for biospecimen metadata based on species

* Update `check_cols_biospecimen()` to have a template match for `human` and set to same template as `general`

* Lint and style issues

* Increment package version