Update study name if choose `No` for study existance
If the user looked through the study names before clicking on "No", then the name that was stopped on will be the study name. This is not necessarily true. If the user then clicks on `No` and does not enter a study name into the textbox before uploading, the study name annotation will be incorrect. Reset the study name back to default of "" any time `No` is selected.