Reset study name if `No` selected
Fixes problem that when the user selects any name from the dropdown box and then selects `No` for study existence. If the user doesn't enter in a study name before uploading data, the wrong study name would have been put on the annotations. This update resets the name to the default of "".