Disable & enable UI elements based on study name
Disable the UI elements if the study name is removed. If the user accidentally selects "" (the first option), then the UI becomes enabled. Made it so the UI elements would disable if the study name is "". This also helps in case the user has chosen a study name from the dropdown (by accident or purpose), but then selects `No` for the study existing without entering any study name in the textbox. The UI elements will simply disable until a name is given.