Check complete cols (#119)
* Update required column checks

Add in the column names that should be complete.

* Improve `success_msg` by making positive

* Fix logic in `check_cols_complete`

`check_cols_complete` now returns `check_fail` if there is 1+ columns missing in the data that are in `required_cols` and/or if the columns present that are required have missing data.

* Update `check_cols_complete` test

Added test cases to make sure the missing required column causes a `check_fail` and returns the missing column name. One case (res4) only tests for a missing column. The other case (res5) tests that the missing column name is included in the returned data even when the other required column fails the test.

* Lintr fighting with styler

Styler changed where the opening curly brace was in this section. Lintr didn't like that. I went with the Lintr version since it makes sense.

* Change `req_cols` to `req_cols_present`

* Changed `condition back to previous version

Didn't need to change the condition here.

* Update documentation for `check_cols_complete()`

* Add updated .Rd documentation

* Increment package version