Require study name in documentation (#122)
* Updated input value for assay_name to be the correct name

* Revert "Updated input value for assay_name to be the correct name"

This reverts commit 9b93bcac2d7d86c4756f854be8fe0b50fcbcfe4e.

* Fix style on same problem as the last 3 PRs

I messed up the style when adding `stringsAsFactors = FALSE` in a previously merged PR. Now I have to keep fixing it in all these PRs until one is merged. I think I've learned my lesson.

* Add placeholder to study name list

Make the first option in the list be "" so that it's easier to enable/disable based on what the current option is.

* Update study name if choose `No` for study existance

If the user looked through the study names before clicking on "No", then the name that was stopped on will be the study name. This is not necessarily true. If the user then clicks on `No` and does not enter a study name into the textbox before uploading, the study name annotation will be incorrect. Reset the study name back to default of "" any time `No` is selected.

* Enable & disable UI elements based on study name

If study name is anything other than "", then the UI elements should be enabled. If the name changes back to "" (i.e. the user clicked on `No`), then the UI elements should be disabled again.

* Revert "Enable & disable UI elements based on study name"

This reverts commit 62cb5e3be19e932f4b080fbce28e1006fb2b07b3.

* Move the `study_name` check and give error message

Moved the `study_name` check to inside the call to `with_busy_indicator_server`. If `study_name() == ""`, an error message will appear by the button and the documents will not be uploaded.