Add annotations, including study name, to metadata & manifest files (#121)
* Team alteration

Altered the team number and folder to upload to for testing.

* Updated input value for assay_name to be the correct name

* Revert "Updated input value for assay_name to be the correct name"

This reverts commit 9b93bcac2d7d86c4756f854be8fe0b50fcbcfe4e.

* Lint, style, & documentation

Added in all the man pages that `devtools::document()` touched even though git doesn't detect changes.

* Add ui for getting the study name

* Get study name from module & enable other ui components

Don't allow the user to upload anything without getting the study name first. Once the study is given a name, then the ui components can activate.

* Update list of study names to have a "" placeholder

This makes it easier for triggering the ui enabling since we can simply check that `study_name() != ""`.

* Add annotations to file uploads

All files upload with `study` name. All files except manifest upload with the `metadataType`. The assay file uploads with `assay` type.

* Lintr still getting made at styler

* Reset study name if `No` selected

Fixes problem that when the user selects any name from the dropdown box and then selects `No` for study existence. If the user doesn't enter in a study name before uploading data, the wrong study name would have been put on the annotations. This update resets the name to the default of "".

* Disable & enable UI elements based on study name

Disable the UI elements if the study name is removed. If the user accidentally selects "" (the first option), then the UI becomes enabled. Made it so the UI elements would disable if the study name is "". This also helps in case the user has chosen a study name from the dropdown (by accident or purpose), but then selects `No` for the study existing without entering any study name in the textbox. The UI elements will simply disable until a name is given.

* Revert "Disable & enable UI elements based on study name"

This reverts commit c0a204608da7b08cb34263b6113133b40c7437f1.

* Revert "Get study name from module & enable other ui components"

This reverts commit 04d8cda11be358d3810ff157844d05a1b13f53d5.

* Activate & populate study name module

Only do this if user passes the sign in checks.

* Collect study name from module

Didn't originally capture the study name from the module.

* Add validate button; start disabled and enable on signin check passed

Include the button indicator feature, which will help with sending an error message if something hasn't been filled out.

* Move report population to `observeEvent` for `validate_btn`

The validation checks will happen in the background, but the report will not populate the success/warn/fail boxes until after the user clicks on `validate_btn`.

* Move saving of files to Synapse to `observeEvent` for `validate_btn`

Files are not uploaded to Synapse until the user clicks `validate_btn`. Gives error message if the user hasn't given study name or any files.

* Fixed variable names

Since the saving process was moved to a different section of the code, needed to use the `files` reactiveValues set. These have slightly different names than the `fileInput` names. Should have tested this on more than just a manifest file.

* Add species annotation

* Turn off linting for cyclomatic complexity

Will turn back on after modularizing the app.

* Update error message for not having uploaded data yet

Co-Authored-By: Kara Woo <karawoo@users.noreply.github.com>