Fix NA and bidirectional issues in check_ids_match. (#141)
* Remove NA from ID lists after getting the differences

There should be NA values in the manifest for IDs in metadata rows, which
should be ignored.
Removing the NA values after getting the setdiff should result in low
processing times for this.

* Change order of data for check_ids_match with manifest

Bidirectional checks the second set of data against the first set of data. We want to check the individual/specimen IDs against those in the manifest. Switching around makes that work.

* Revert "Change order of data for check_ids_match with manifest"

This reverts commit f511bf55384c0df4d3af72ded39618b2c8ca8159.