Update `is_name_valid()` to check for NULL and skip unnecessary code
Needed to check for a NULL filename; added that.

Realized that the way the function was written, it would do unnecessary checking if the name was NULL or "" since I don't `return(valid)` early. Instead of adding multiple `return` statements, I just made a more elaborate `if-else`. Note that I originally was checking `if (is.null(name) || name == "")` at first, but the tests would fail due to getting back `logical(0)` instead of FALSE. I simply broke these into two and it works now.