Check file names valid (#191)
* Rename check-study-name.R to more generic is_name_valid and update

Update check-study-name.R to be more generic for checking both study names and file names for validity. Update documentation.

* Update tests since changed function name to more generic version

* Update the function name when checking study name

* Add checks for file name validity before saving to synapse

* Remove repetitive check for valid filename from server.

This check is to be done in `save_to_synapse()`

* Add check for valid file name in `save_to_synapse()`

* Add test to check result for NULL filename

* Update `is_name_valid()` to check for NULL and skip unnecessary code

Needed to check for a NULL filename; added that.

Realized that the way the function was written, it would do unnecessary checking if the name was NULL or "" since I don't `return(valid)` early. Instead of adding multiple `return` statements, I just made a more elaborate `if-else`. Note that I originally was checking `if (is.null(name) || name == "")` at first, but the tests would fail due to getting back `logical(0)` instead of FALSE. I simply broke these into two and it works now.

* Update R/check-name.R documentation

Co-Authored-By: Kara Woo <karawoo@users.noreply.github.com>

* Update R/check-name.R to be marked as internal

Co-Authored-By: Kara Woo <karawoo@users.noreply.github.com>

* Update call to `is_name_valid()` to use name from `input_file`

* Regenerate .Rd documentation

* Update R/utils.R

Co-Authored-By: Kara Woo <karawoo@users.noreply.github.com>