Move instructions text into a function
Added `upload_docs_instruct_text()` that creates the html taglist for the upload documentation instructions. This takes the links from the config file for human and animal examples, as well as the acknowledgements example. If `study_link_ref` is an empty string, then the acknowledgement instructions are left out; otherwise, they are included with a link to examples.