Add link and text for study acknowledgments to document upload (#200)
* Add link and text for study acknowledgments to `upload_documents_ui()`

Add text about what acknowledgments are and how to upload them in the upload documents instructions. The added link allows for customizing the example shown for acknowlegements/references.

* Change default study_link_ref

Move the amp-ad link to the amp-ad section, and make default be an empty string. An empty string will make the acknowledgement section not appear.

* Move instructions text into a function

Added `upload_docs_instruct_text()` that creates the html taglist for the upload documentation instructions. This takes the links from the config file for human and animal examples, as well as the acknowledgements example. If `study_link_ref` is an empty string, then the acknowledgement instructions are left out; otherwise, they are included with a link to examples.

* Fix grammar error.

Co-Authored-By: Kara Woo <karawoo@users.noreply.github.com>

* Make consistent with instructions.

Co-Authored-By: Kara Woo <karawoo@users.noreply.github.com>