Remove AMP-AD team from default settings
AMP-AD doesn't have permissions to create folders in the validation-sandbox project, which causes the app to crash if they log in while the config is set to default. Ultimately, it would be better to have a default view that shows it's not the AMP-AD version. This is a quick solution, but will not result in much less confusion.